Week 6 - 9/SAM_PROJECT_TWO_WORKS!/requirements.txt
fastapi
uvicorn
mangum
python-dotenv
langchain-openai
boto3
